pyproject.toml
[project]
name = "fflow"
version = "0.1.0"
description = "ForesightFlow — real-time informed-trading detection for Polymarket"
requires-python = ">=3.12"
dependencies = [
    "httpx[http2]>=0.27",
    "gql[httpx]>=3.5",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.13",
    "asyncpg>=0.29",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "typer>=0.12",
    "structlog>=24.0",
    # Task 02
    "pandas>=2.2",
    "beautifulsoup4>=4.12",
    "lxml>=5.0",
    "nltk>=3.8",
    "anthropic>=0.25",
    "pyyaml>=6.0",
    "google-cloud-bigquery>=3.0; extra == 'gdelt'",

    "pyarrow>=24.0.0",
    "scipy>=1.17.1",
]

[project.optional-dependencies]
gdelt = ["google-cloud-bigquery>=3.0"]

[project.scripts]
fflow = "fflow.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "aiosqlite>=0.22.1",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-recording>=0.13",
    "vcrpy>=6.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
